wip, why does picking not work on our custom meshes?
